Move SuiteFactory to another package (#50)
